chore(tests): updated test files per functionality
