+++ src/deps.ts
@@ -1,1 +1,1 @@
-export { existsSync } from 'jsr:@std/fs@1.0.21';
+export { existsSync } from 'jsr:@std/fs@1.0.23';
