+++ tests/deps.ts
@@ -13,8 +13,8 @@ export {
     assertArrayIncludes,
     assertStrictEquals,
     assertNotStrictEquals
-} from 'jsr:@std/assert@1.0.16';
-export { assertSpyCalls, spy } from 'jsr:@std/testing@1.0.16/mock';
+} from 'jsr:@std/assert@1.0.19';
+export { assertSpyCalls, spy } from 'jsr:@std/testing@1.0.17/mock';
 export { 
     describe, 
     it, 
@@ -22,7 +22,7 @@ export {
     beforeEach, 
     beforeAll, 
     afterEach 
-} from 'jsr:@std/testing@1.0.16/bdd';
-export { delay } from 'jsr:@std/async@1.0.16/delay';
+} from 'jsr:@std/testing@1.0.17/bdd';
+export { delay } from 'jsr:@std/async@1.2.0/delay';
 export { load } from 'jsr:@std/dotenv@0.225.6';
 export * as mf from 'https://deno.land/x/mock_fetch@0.3.0/mod.ts';
